refactor: events (#2885)
